// src/types/bailleur.ts
export interface BailleurStats {
  total_biens: number
  biens_libres: number
  biens_occupes: number
  taux_occupation: number
  revenus_mois: number
  revenus_encaisses: number
  impayes: number
  montant_impayes: number
  loyers_confirmes: number
  signalements: number
  baux_renouveler: number
}